[future: Rivsen/wand#1]: split sub commands
