+++ Cargo.toml
@@ -15,3 +15,4 @@ prettytable-rs = "0.8.0"
 serde = { version = "1.0.126", features = ["derive"] }
 serde_json = "1.0.64"
 tera = "1.11.0"
+actix-web = "3.3.2"
